Database_type config to Split up sql queries by engine type (#7934)
